+++ app/page.tsx
@@ -1,8 +1,7 @@
 "use client";
 
-import RoutePage from "./components/RoutePage";
+import RoutePage from "./components/pages/RoutePage";
 import { ROUTES } from "./components/layout/SideBar";
-import "./globals.css";
 
 export default function Home() {
   return <RoutePage title="Apps" routes={ROUTES.slice(1)} />;
