+++ app/specialOccasions/2024/page.tsx
@@ -1,7 +1,7 @@
-import RoutePage from "../../components/RoutePage";
+import RoutePage from "../../components/pages/RoutePage";
 import { FiHeart } from "react-icons/fi";
 import { IoRose } from "react-icons/io5";
-import { Route } from "../../components/RoutePage";
+import { Route } from "../../components/pages/RoutePage";
 
 const ROUTES: Route[] = [
   {
